+++ src/pl/plajer/buildbattle3/handlers/SignManager.java
@@ -176,11 +176,12 @@ public class SignManager implements Listener {
   private void updateSignScheduler() {
     Bukkit.getScheduler().runTaskTimer(plugin, () -> {
       for (Sign s : loadedSigns.keySet()) {
+        Block block = s.getBlock();
         for (int i = 0; i < signLines.size(); i++) {
           s.setLine(i, formatSign(signLines.get(i), loadedSigns.get(s)));
-          if (plugin.getConfig().getBoolean("Signs-Block-States-Enabled")) {
-            if (s.getType() == Material.SIGN_POST || s.getType() == Material.WALL_SIGN) {
-              Block behind = s.getBlock().getRelative(((org.bukkit.material.Sign) s.getData()).getAttachedFace());
+          if (plugin.getConfig().getBoolean("Signs-Block-States-Enabled", true)) {
+            if (block.getType() == Material.SIGN_POST || block.getType() == Material.WALL_SIGN) {
+              Block behind = block.getRelative(((org.bukkit.material.Sign) s.getData()).getAttachedFace());
               behind.setType(Material.STAINED_GLASS);
               switch (loadedSigns.get(s).getArenaState()) {
                 case WAITING_FOR_PLAYERS: